requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
yagmail==0.15.280
pandas==2.1.1
python-dotenv==1.0.0